Refactor EditProfileModal into separate page.
